shared/requirements.txt
grpcio>=1.63.2,<2
grpcio-tools>=1.63.2,<2
protobuf>=4.25.3,<6
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
nats-py==2.6.0
etcd3-py==0.1.6
prometheus-client==0.20.0
opentelemetry-api>=1.30.0,<2
opentelemetry-sdk>=1.30.0,<2
opentelemetry-exporter-otlp-proto-grpc>=1.30.0,<2
opentelemetry-instrumentation-grpc>=0.51b0
opentelemetry-instrumentation-fastapi>=0.51b0
pydantic==2.6.4
